7/chore - корректировка типа данных для времени в мс

+++ src/main/java/ru/skypro/homework/dto/CommentDTO.java
@@ -10,7 +10,7 @@ public class CommentDTO {
     private Integer author;
     private String authorImage;
     private String authorFirstName;
-    private Integer createdAt;
+    private Long createdAt;
     //id for comment
     private Integer pk;
